# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Assessment System API",
    "description": "A comprehensive API for managing assessments, submissions, and reviews in a sustainability assessment system.\n\n## Authentication\nAll endpoints require JWT authentication. Include the token in the Authorization header:\n`Authorization: Bearer <your-jwt-token>`\n\n## Versioning\nThe API supports response versioning to track changes and resolve conflicts.\n",
    "contact": {
      "name": "Assessment System Team",
      "email": "support@assessment-system.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/submissions": {
      "get": {
        "tags": ["Submissions"],
        "summary": "List user submissions",
        "description": "Get user's submitted assessments",
        "operationId": "list_user_submissions",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by review status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending_review",
                "under_review",
                "approved",
                "rejected",
                "revision_requested"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user submissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/submissions/{submission_id}": {
      "get": {
        "tags": ["Submissions"],
        "summary": "Get submission",
        "description": "Get a specific submission",
        "operationId": "get_submission",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "description": "Submission ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submission details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionDetailResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Submissions"],
        "summary": "Delete submission",
        "description": "Delete a submission (Organization admins only)",
        "operationId": "delete_submission",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "description": "Submission ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Submission deleted successfully"
          },
          "400": {
            "description": "Bad Request - Only organization admins can delete submissions or submission not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Submission": {
        "required": [
          "assessment_id",
          "assessment_name",
          "content",
          "review_status",
          "submission_id",
          "submitted_at",
          "user_id"
        ],
        "type": "object",
        "properties": {
          "submission_id": {
            "type": "string",
            "format": "uuid"
          },
          "assessment_id": {
            "type": "string",
            "format": "uuid"
          },
          "assessment_name": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/Submission_content"
          },
          "review_status": {
            "type": "string",
            "enum": [
              "pending_review",
              "under_review",
              "approved",
              "rejected",
              "revision_requested"
            ]
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "SubmissionListResponse": {
        "required": ["submissions"],
        "type": "object",
        "properties": {
          "submissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Submission"
            }
          }
        }
      },
...
      "SubmissionDetailResponse": {
        "required": ["current_review", "submission"],
        "type": "object",
        "properties": {
          "submission": {
            "$ref": "#/components/schemas/Submission"
          },
          "current_review": {
            "$ref": "#/components/schemas/Review"
          }
        }
      },
...
      "Review": {
        "required": [
          "created_at",
          "review_id",
          "reviewer_id",
          "status",
          "submission_id"
        ],
        "type": "object",
        "properties": {
          "review_id": {
            "type": "string",
            "format": "uuid"
          },
          "submission_id": {
            "type": "string",
            "format": "uuid"
          },
          "reviewer_id": {
            "type": "string"
          },
          "reviewer_email": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed"]
          },
          "decision": {
            "type": "string",
            "nullable": true,
            "enum": ["approved", "rejected", "revision_requested"]
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "FileMetadata": {
        "required": ["file_id", "metadata"],
        "type": "object",
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {
            "type": "object",
            "description": "File metadata including filename, size, content type, etc."
          }
        }
      },
...
      "Error": {
        "required": ["error", "message"],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "Submission_content_assessment": {
        "type": "object",
        "properties": {
          "assessment_id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "language": {
            "type": "string",
            "example": "en"
          }
        }
      },
      "Submission_content_responses": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "format": "uuid"
          },
          "response": {
            "type": "string",
            "example": "Our organization has a comprehensive sustainability policy"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileMetadata"
            }
          }
        }
      },
      "Submission_content": {
        "type": "object",
        "properties": {
          "assessment": {
            "$ref": "#/components/schemas/Submission_content_assessment"
          },
          "responses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Submission_content_responses"
            }
          }
        },
        "description": "Immutable snapshot of assessment and all responses"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or category",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message",
            "example": "Invalid organization data provided"
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}